requirements.txt
PySide6==6.5.0
opencv-python==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
requests==2.29.0
tqdm>=4.65.0
rembg>=2.0.30 